requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langgraph-checkpoint-postgres>=2.0.0
langchain-anthropic>=0.3.0
langsmith>=0.1.0
anthropic>=0.40.0
openai>=1.40.0
google-generativeai>=0.8.0
cohere>=5.0.0
voyageai>=0.3.0
sentence-transformers>=3.0.0
qdrant-client>=1.9.0
psycopg2-binary>=2.9.0
pgvector>=0.3.0
pinecone-client>=3.0.0
chromadb>=0.5.0
celery>=5.4.0
redis>=5.0.0
kombu>=5.3.0
watchdog>=4.0.0
networkx>=3.3
graspologic>=3.3.0
pymupdf>=1.24.0
python-docx>=1.1.0
opencv-python-headless>=4.10.0
faster-whisper>=1.0.0
Pillow>=10.0.0
python-magic>=0.4.27
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
httpx>=0.27.0
slowapi>=0.1.9
prometheus-client>=0.20.0
pyyaml>=6.0.0
numpy>=1.26.0
